refactor(ds-global-form): rename choices components
Rename the two choice-group components to better names:

- SimpleChoicesField (plain radio/checkbox list) -> ChoicesField
- ChoicesField (rich, freely-styled cards)        -> RichChoicesField

Covers the presentational components (SimpleChoices -> Choices, Choices ->
RichChoices), prop types, folders/files, index barrels, and the Field pattern's
discriminated union + switch. Public-facing identifiers move too:

- inputType:  'simple-choices' -> 'choices',  'choices' -> 'rich-choices'
- CSS class:  .ds.form-simple-choices -> .ds.form-choices,
              .ds.form-choices -> .ds.form-rich-choices
- custom props: --simple-choices-columns -> --choices-columns,
                --choices-span -> --rich-choices-span

BREAKING CHANGE: the 'simple-choices'/'choices' inputType strings and the
.ds.form-* class names change, so consumers selecting these via <Field> or
theming the classes must update.

tsc + biome + full suite (236) pass; storybook builds.
