+++ packages/react/ds-global-form/src/lib/component/ChoicesField/Choices.tsx
@@ -1,3 +1,4 @@
+/* @canonical/generator-ds 0.9.0-experimental.9 */
 import type React from "react";
 import { Option } from "./common/index.js";
 import type { ChoicesPresentationProps } from "./types.js";
@@ -6,7 +7,7 @@ import "./styles.css";
 const componentCssClassName = "ds form-choices";
 
 /**
- * Hidden-input fieldset with freely-styled labels (cards, icons, etc.).
+ * Plain-label option group (radios or checkboxes) with inline/stacked layout.
  * Controlled and form-agnostic: the selected value(s) flow in via `value` and
  * out via `onChange` (a single value for radios, an array for checkboxes).
  * Each option is rendered by the `Option` subcomponent (common/Option).
@@ -19,17 +20,31 @@ export const Choices = ({
   name,
   isMultiple = false,
   disabled = false,
+  layout = "inline",
+  columns,
   options,
   value,
   onChange,
 }: ChoicesPresentationProps): React.ReactElement => {
   const type = isMultiple ? "checkbox" : "radio";
 
+  // In the "columns" layout the column count drives a CSS grid; the variable is
+  // unused (and so left unset) in the other layouts.
+  const layoutStyle =
+    layout === "columns" && columns
+      ? ({
+          ...style,
+          "--choices-columns": columns,
+        } as React.CSSProperties)
+      : style;
+
   return (
     <fieldset
       id={id}
-      style={style}
-      className={[componentCssClassName, className].filter(Boolean).join(" ")}
+      style={layoutStyle}
+      className={[componentCssClassName, layout, className]
+        .filter(Boolean)
+        .join(" ")}
     >
       {options.map((option) => {
         const checked = isMultiple
