+++ packages/react/ds-global-form/src/lib/component/ChoicesField/ChoicesField.tests.tsx
@@ -1,7 +1,7 @@
 import { fireEvent, screen } from "@testing-library/react";
 import { describe, expect, it } from "vitest";
 import { renderWithForm } from "../../../testing/renderWithForm.js";
-import { SimpleChoicesField } from "./index.js";
+import { ChoicesField } from "./index.js";
 
 const options = [
   { label: "Red", value: "red" },
@@ -9,30 +9,26 @@ const options = [
   { label: "Green", value: "green" },
 ];
 
-describe("SimpleChoicesField", () => {
+describe("ChoicesField", () => {
   it("renders radio inputs by default", () => {
-    renderWithForm(<SimpleChoicesField name="color" options={options} />);
+    renderWithForm(<ChoicesField name="color" options={options} />);
     expect(screen.getAllByRole("radio")).toHaveLength(3);
   });
 
   it("renders checkbox inputs when isMultiple", () => {
-    renderWithForm(
-      <SimpleChoicesField name="colors" options={options} isMultiple />,
-    );
+    renderWithForm(<ChoicesField name="colors" options={options} isMultiple />);
     expect(screen.getAllByRole("checkbox")).toHaveLength(3);
   });
 
   it("renders labels for each option", () => {
-    renderWithForm(<SimpleChoicesField name="color" options={options} />);
+    renderWithForm(<ChoicesField name="color" options={options} />);
     expect(screen.getByText("Red")).toBeInTheDocument();
     expect(screen.getByText("Blue")).toBeInTheDocument();
     expect(screen.getByText("Green")).toBeInTheDocument();
   });
 
   it("supports disabled state on all options", () => {
-    renderWithForm(
-      <SimpleChoicesField name="color" options={options} disabled />,
-    );
+    renderWithForm(<ChoicesField name="color" options={options} disabled />);
     for (const radio of screen.getAllByRole("radio")) {
       expect(radio).toBeDisabled();
     }
@@ -43,13 +39,13 @@ describe("SimpleChoicesField", () => {
       ...options,
       { label: "Disabled", value: "disabled", disabled: true },
     ];
-    renderWithForm(<SimpleChoicesField name="color" options={opts} />);
+    renderWithForm(<ChoicesField name="color" options={opts} />);
     expect(screen.getAllByRole("radio")).toHaveLength(4);
     expect(screen.getAllByRole("radio")[3]).toBeDisabled();
   });
 
   it("selects a radio option on click", () => {
-    renderWithForm(<SimpleChoicesField name="color" options={options} />);
+    renderWithForm(<ChoicesField name="color" options={options} />);
     const radios = screen.getAllByRole("radio");
     fireEvent.click(radios[1]);
     expect(radios[1]).toBeChecked();
