+++ packages/react/ds-global-form/src/lib/component/ChoicesField/ChoicesField.tsx
@@ -4,8 +4,8 @@ import { Choices } from "./Choices.js";
 import type { ChoicesFieldProps } from "./types.js";
 
 /**
- * Choices bound to react-hook-form (controlled), wrapped with field chrome.
- * The label is mocked into the fieldset legend, as the per-option labels are
+ * Choices bound to react-hook-form (controlled), wrapped with field
+ * chrome. The label is mocked into the fieldset legend, as the "true" label is
  * rendered by the individual options.
  */
 export default withWrapper<ChoicesFieldProps>(
