+++ packages/react/ds-global-form/src/lib/component/ChoicesField/common/Option/Option.stories.tsx
@@ -9,7 +9,7 @@ const meta: Meta<typeof Option> = {
   tags: ["autodocs"],
   decorators: [
     (Story) => (
-      <fieldset className="ds form-choices" style={{ border: "none" }}>
+      <fieldset className="ds form-choices inline" style={{ border: "none" }}>
         <Story />
       </fieldset>
     ),
