+++ packages/react/ds-global-form/src/lib/component/ChoicesField/common/Option/types.ts
@@ -1,6 +1,6 @@
 // Generated by @canonical/summon-component:react v0.1.0
 
-import type { ChoiceOption } from "../../types.js";
+import type { Option as ChoiceOption } from "#lib/subcomponent/types.js";
 
 export interface OptionProps {
   /** Field name shared by every option input in the group */
