+++ packages/react/ds-global-form/src/lib/component/ChoicesField/styles.css
@@ -1,93 +1,152 @@
 .ds.form-choices {
   border: none;
   padding: 0;
+  padding-inline: var(--form-input-padding-inline);
   margin: 0;
-  display: grid;
-  grid-template-columns: subgrid;
-  grid-column: 1 / -1;
-  row-gap: var(--form-group-gap);
+  display: flex;
+  flex-wrap: wrap;
+  gap: 0 var(--container-gap-loose);
+
+  /* Stacked: one option per line */
+  &.stacked {
+    flex-direction: column;
+    flex-wrap: nowrap;
+  }
+
+  /* Columns: a grid of equal-width columns. Each option fills its cell, so the
+   * option WIDTH is column-based rather than content-based. The column count
+   * comes from `--choices-columns` (set from the `columns` prop). */
+  &.columns {
+    display: grid;
+    grid-template-columns: repeat(var(--choices-columns, 2), 1fr);
+    gap: var(--form-group-gap) var(--container-gap-loose);
+  }
 
   /* Each option (common/Option, `.ds.option`) is styled here, scoped to this
    * group so the generic class cannot collide with options elsewhere. */
   & > .ds.option {
-    grid-column: var(--choices-span, auto);
-    position: relative;
+    direction: inherit;
     display: flex;
+    align-items: baseline;
+    gap: var(--space-100);
+    cursor: pointer;
 
-    /* Hide the native input */
-    & > input {
-      position: absolute;
-      width: 1px;
-      height: 1px;
-      overflow: hidden;
-      clip: rect(0, 0, 0, 0);
-      white-space: nowrap;
-      border: 0;
-    }
-
-    /* Label is the visible card surface */
-    & > label {
-      display: flex;
-      flex-direction: column;
-      flex: 1;
-      box-sizing: border-box;
-      gap: var(--form-group-gap);
-      padding-inline: var(--form-input-padding-inline);
-      /* Block padding: chrome padding + baseline nudges, minus border */
-      padding-block-start: calc(
-        var(--form-input-padding-block) +
-        var(--start-nudge, 0px) -
-        var(--form-input-border-width)
-      );
-      padding-block-end: calc(
-        var(--form-input-padding-block) +
-        var(--end-nudge, 0px) -
-        var(--form-input-border-width)
-      );
-      border: var(--form-input-border-width) solid
-        var(--form-input-border-color);
-      border-radius: var(--form-input-border-radius);
-      background-color: var(--form-input-background);
-      color: var(--form-input-color);
-      cursor: pointer;
-      transition:
-        border-color 150ms ease,
-        background-color 150ms ease,
-        box-shadow 150ms ease;
-      user-select: none;
-    }
-
-    /* Hover */
-    & > input:not(:disabled) + label:hover {
-      background-color: var(
-        --hover--color-foreground-input,
-        var(--color-foreground-input-hover)
-      );
-    }
-
-    /* Focus */
-    & > input:focus-visible + label {
-      outline: none;
-      border-color: var(--form-input-border-color-focus);
-      box-shadow: 0 0 0 1px var(--form-focus-ring-color);
-    }
-
-    /* Checked / selected */
-    & > input:checked + label {
-      border-color: var(--form-input-border-color-focus);
-      background-color: var(
-        --hover--color-foreground-input,
-        var(--color-foreground-input-hover)
-      );
-      box-shadow: 0 0 0 1px var(--form-input-border-color-focus);
-    }
-
-    /* Disabled */
-    & > input:disabled + label {
-      border-color: var(--form-input-border-color-disabled);
-      background-color: var(--form-input-background-disabled);
-      color: var(--form-input-color-disabled);
+    &.disabled {
       cursor: not-allowed;
+      opacity: 0.5;
+    }
+
+    & > input[type="radio"] {
+      appearance: none;
+      box-sizing: border-box;
+      width: calc(var(--space-baseline) * 2); /* 16px — 2bU */
+      height: calc(var(--space-baseline) * 2); /* 16px — 2bU */
+      border: var(--dimension-stroke-thickness-medium) solid var(--color-border);
+      border-radius: 50%;
+      background-color: var(--color-foreground-radio-unselected);
+      cursor: inherit;
+      transition:
+        background-color 150ms ease,
+        border-color 150ms ease;
+
+      &:checked {
+        border-color: var(--color-foreground-radio-selected);
+        background: radial-gradient(
+          var(--color-foreground-radio-selected) 40%,
+          var(--color-foreground-radio-unselected) 41%
+        );
+      }
+
+      &:hover:not(:disabled):not(:checked) {
+        background-color: var(--color-foreground-radio-unselected-hover);
+      }
+
+      &:checked:hover:not(:disabled) {
+        filter: brightness(0.9);
+      }
+
+      &:focus-visible {
+        outline: none;
+        border-color: var(--color-focusRing);
+        box-shadow: 0 0 0 1px var(--color-focusRing);
+      }
+
+      &:disabled {
+        border-color: var(--color-border-disabled);
+        background-color: var(--color-foreground-radio-unselected-disabled);
+      }
+    }
+
+    & > input[type="checkbox"] {
+      appearance: none;
+      box-sizing: border-box;
+      position: relative;
+      width: calc(var(--space-baseline) * 2); /* 16px — 2bU */
+      height: calc(var(--space-baseline) * 2); /* 16px — 2bU */
+      border: var(--dimension-stroke-thickness-medium) solid var(--color-border);
+      border-radius: var(--dimension-radius-small);
+      background-color: var(--color-foreground-checkbox-unselected);
+      cursor: inherit;
+      transition:
+        background-color 150ms ease,
+        border-color 150ms ease;
+
+      /* Glyph as a masked pseudo-element so it takes the checkmark token
+       * (see CheckboxInput); the colour+mask live together under :checked so
+       * the unchecked box never paints a solid square. */
+      &::before {
+        content: "";
+        position: absolute;
+        inset: 0;
+        mask-repeat: no-repeat;
+        mask-position: center;
+        mask-size: 75%;
+        -webkit-mask-repeat: no-repeat;
+        -webkit-mask-position: center;
+        -webkit-mask-size: 75%;
+      }
+
+      &:checked {
+        background: var(--color-foreground-checkbox-selected);
+        border-color: var(--color-foreground-checkbox-selected);
+
+        &::before {
+          background-color: var(--color-foreground-checkbox-checkmark);
+          mask-image: url("/icons/checkmark.svg#checkmark");
+          -webkit-mask-image: url("/icons/checkmark.svg#checkmark");
+        }
+      }
+
+      &:hover:not(:disabled):not(:checked) {
+        background-color: var(--color-foreground-checkbox-unselected-hover);
+      }
+
+      &:checked:hover:not(:disabled) {
+        filter: brightness(0.9);
+      }
+
+      &:focus-visible {
+        outline: none;
+        border-color: var(--color-focusRing);
+        box-shadow: 0 0 0 1px var(--color-focusRing);
+      }
+
+      &:disabled {
+        border-color: var(--color-border-disabled);
+        background-color: var(--color-foreground-checkbox-unselected-disabled);
+      }
+
+      /* A checked+disabled box must keep the SELECTED (not unselected) disabled
+       * fill — the generic `:disabled` above would otherwise leave the checkmark
+       * on the unselected-disabled background. Matches CheckboxInput. */
+      &:checked:disabled {
+        background: var(--color-foreground-checkbox-selected-disabled);
+        border-color: var(--color-foreground-checkbox-selected-disabled);
+
+        &::before {
+          background-color: var(--color-foreground-checkbox-checkmark-disabled);
+        }
+      }
     }
   }
 }
