+++ packages/react/ds-global-form/src/lib/component/RichChoicesField/RichChoices.ssr.tests.tsx
@@ -1,6 +1,6 @@
 import { renderToString } from "react-dom/server";
 import { describe, expect, it } from "vitest";
-import { SimpleChoices } from "./SimpleChoices.js";
+import { RichChoices } from "./RichChoices.js";
 
 // Proves the presentational input renders to static HTML with no client
 // runtime / form context — the server-rendered floor for progressive
@@ -10,13 +10,11 @@ const options = [
   { label: "Blue", value: "blue" },
 ];
 
-describe("SimpleChoices (SSR)", () => {
+describe("RichChoices (SSR)", () => {
   it("renders the fieldset and option inputs to static HTML", () => {
-    const html = renderToString(
-      <SimpleChoices name="color" options={options} />,
-    );
+    const html = renderToString(<RichChoices name="color" options={options} />);
     expect(html).toContain("<fieldset");
-    expect(html).toContain("ds form-simple-choices");
+    expect(html).toContain("ds form-rich-choices");
     expect(html).toContain("<input");
     expect(html).toContain('type="radio"');
     expect(html).toContain('name="color"');
@@ -25,7 +23,7 @@ describe("SimpleChoices (SSR)", () => {
 
   it("renders checkbox inputs for the multiple variant", () => {
     const html = renderToString(
-      <SimpleChoices name="colors" options={options} isMultiple />,
+      <RichChoices name="colors" options={options} isMultiple />,
     );
     expect(html).toContain('type="checkbox"');
   });
