+++ packages/react/ds-global-form/src/lib/component/RichChoicesField/RichChoices.tsx
@@ -1,50 +1,35 @@
-/* @canonical/generator-ds 0.9.0-experimental.9 */
 import type React from "react";
 import { Option } from "./common/index.js";
-import type { SimpleChoicesPresentationProps } from "./types.js";
+import type { RichChoicesPresentationProps } from "./types.js";
 import "./styles.css";
 
-const componentCssClassName = "ds form-simple-choices";
+const componentCssClassName = "ds form-rich-choices";
 
 /**
- * Plain-label option group (radios or checkboxes) with inline/stacked layout.
+ * Hidden-input fieldset with freely-styled labels (cards, icons, etc.).
  * Controlled and form-agnostic: the selected value(s) flow in via `value` and
  * out via `onChange` (a single value for radios, an array for checkboxes).
  * Each option is rendered by the `Option` subcomponent (common/Option).
- * @returns {React.ReactElement} - Rendered SimpleChoices
+ * @returns {React.ReactElement} - Rendered RichChoices
  */
-export const SimpleChoices = ({
+export const RichChoices = ({
   id,
   className,
   style,
   name,
   isMultiple = false,
   disabled = false,
-  layout = "inline",
-  columns,
   options,
   value,
   onChange,
-}: SimpleChoicesPresentationProps): React.ReactElement => {
+}: RichChoicesPresentationProps): React.ReactElement => {
   const type = isMultiple ? "checkbox" : "radio";
-
-  // In the "columns" layout the column count drives a CSS grid; the variable is
-  // unused (and so left unset) in the other layouts.
-  const layoutStyle =
-    layout === "columns" && columns
-      ? ({
-          ...style,
-          "--simple-choices-columns": columns,
-        } as React.CSSProperties)
-      : style;
 
   return (
     <fieldset
       id={id}
-      style={layoutStyle}
-      className={[componentCssClassName, layout, className]
-        .filter(Boolean)
-        .join(" ")}
+      style={style}
+      className={[componentCssClassName, className].filter(Boolean).join(" ")}
     >
       {options.map((option) => {
         const checked = isMultiple
@@ -78,4 +63,4 @@ export const SimpleChoices = ({
   );
 };
 
-export default SimpleChoices;
+export default RichChoices;
