+++ packages/react/ds-global-form/src/lib/component/RichChoicesField/common/Option/index.ts
