+++ packages/react/ds-global-form/src/lib/component/RichChoicesField/common/Option/types.ts
@@ -1,6 +1,6 @@
 // Generated by @canonical/summon-component:react v0.1.0
 
-import type { Option as ChoiceOption } from "#lib/subcomponent/types.js";
+import type { RichChoiceOption } from "../../types.js";
 
 export interface OptionProps {
   /** Field name shared by every option input in the group */
@@ -8,7 +8,7 @@ export interface OptionProps {
   /** Native input type backing the option (radios vs checkboxes) */
   type: "checkbox" | "radio";
   /** The option to render */
-  option: ChoiceOption;
+  option: RichChoiceOption;
   /** Whether this option is currently selected */
   checked: boolean;
   /** Whether this option is disabled */
