+++ packages/react/ds-global-form/src/lib/component/RichChoicesField/index.ts
@@ -1,0 +1,2 @@
+export { default as RichChoicesField } from "./RichChoicesField.js";
+export type { RichChoiceOption, RichChoicesFieldProps } from "./types.js";
