--- packages/react/ds-global-form/src/lib/component/SimpleChoicesField/index.ts
@@ -1,2 +1,0 @@
-export { default as SimpleChoicesField } from "./SimpleChoicesField.js";
-export type { SimpleChoicesFieldProps } from "./types.js";
