+++ packages/react/ds-global-form/src/lib/component/index.ts
@@ -12,8 +12,8 @@ export * from "./NumberField/index.js";
 export * from "./PasswordField/index.js";
 export * from "./PhoneField/index.js";
 export * from "./RangeField/index.js";
+export * from "./RichChoicesField/index.js";
 export * from "./SelectField/index.js";
-export * from "./SimpleChoicesField/index.js";
 export * from "./TextareaField/index.js";
 export * from "./TextField/index.js";
 export * from "./TimeField/index.js";
