+++ packages/react/ds-global-form/src/storybook/fixtures.fields.ts
@@ -58,17 +58,17 @@ export const base = [
     placeholder: "Type your message here",
   },
   {
-    name: "simple_choices_single",
-    inputType: "simple-choices",
+    name: "choices_single",
+    inputType: "choices",
     label: "Choose a continent",
-    description: "This is a simple choices field",
+    description: "This is a choices field",
     options: options.continents,
   },
   {
-    name: "simple_choices_multiple",
-    inputType: "simple-choices",
+    name: "choices_multiple",
+    inputType: "choices",
     label: "Choose multiple continents",
-    description: "This is a simple choices field",
+    description: "This is a choices field",
     isMultiple: true,
     options: options.continents,
   },
